% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_df.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert a Tensor to a Data Frame}
\usage{
convert_to_df(item)
}
\arguments{
\item{item}{A PyTorch tensor, a list of tensors, or another R object.}
}
\value{
A data frame or a list of data frames.
}
\description{
This function converts a PyTorch tensor (or a list of tensors) into an R data frame.
If the tensor has at least two columns, the last column is assumed to be a class label.
}
\examples{
\dontrun{
  torch <- reticulate::import("torch")
  tensor <- torch$randn(c(10, 3))  # Random 10x3 tensor
  df <- convert_to_df(tensor)
  print(df)
}
}
