% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tensor.R
\name{convert_to_tensor}
\alias{convert_to_tensor}
\title{Convert a List or Data Frame to a PyTorch Tensor}
\usage{
convert_to_tensor(lst)
}
\arguments{
\item{lst}{A data frame, list of data frames, or numeric matrix to convert into a tensor.}
}
\value{
A PyTorch tensor or a list of tensors.
}
\description{
This function converts an R list or data frame into a PyTorch tensor using `torch$from_numpy()`.
}
\examples{
\dontrun{
  torch <- reticulate::import("torch")
  df <- data.frame(A = rnorm(10), B = rnorm(10))
  tensor <- convert_to_tensor(df)
  print(tensor)
}
}
